#3, #6: Verified I2C Bus Multiplexor. Verified sensor cluster base functions
As per issue 3, the I2C multiplexer appears to be responding to
switching commands and successfully routing I2C commands through to the
correct module.

As per issue 6, the sensor cluster class and its respective methods
appear to be working as intended. Creating individual SensorCluster
modules with all of the needed information to carry out data acquisition
provides a clean and elegant way to interface to the web interface, as a
caller would be able to run `updateAllSensors()` followed by
`saveAllSensors()`

I have yet to test the actual interface between a database, as I need to
put in a bit more work to understand the database's exposed methods.