#8 Initial prototype of PCB Added
The files added as part of this commit are all associated with the first
prototype of the PCB. I'll be removing a lot of these files later on as
I figure out exactly what files are needed in order for someone else to
view and generate the files needed to print their own board.

A 3D rendering of the board has also been added.