#11: Added Development Controls Interface
As per previous commit, I noticed a slight mistake with the last file.

`controls.py` serves as a basic brainstorming controls interface that
will be expanded and modified a lot in the next couple of days.

Working on this script helped me to realize that using an IO expander is
probably the way to go; The GPIO header was starting to get pretty
crowded when I was going through all of the different control modules.