Restructured Sensor Model
This new sensor model structure hides each sensor cluster behind a
single node head (in this case, the node head is an I2C multiplexer).
This eliminates the need for individual sensor addresses and further
generalizes the models. This also abstracts more hardware away from the
python script and thus will make it easier for a user to add sensors.