#11: Created control model, added support for IO expander
The control model contained in this commit should make interfacing to
the control modules a little bit easier by abstracting a lot of low
level register manipulation into the model.

Each control cluster will need to be instantiated with an ID. It will
also receive a SensorCluster object (of the same ID) and a list of ideal
values for the given plant.

Methods within the control model exist that compare these ideal and
sensed values and then build a mask that can then be stored into the IO
expander once all of the control unit pins have been configured.

Utilities have also been added to support the MCP IO expander. These can
be seen in `i2c_utility.py`