#11 Verified physical functionality - added test cases
Additional test cases have been added to the controltest.py script in
order to test and verify physical function through a series of LEDs
acting as a proxy for the control modules later on.

i2c_utility.py was modified slightly to fix an error with setting the
pin IO direction.