Close #5: Added soil moisture sensing
Soil moisture sensors now have full support. I'm closing this issue as
we may be removing pH sensing from the project for now, though space
will exist on the ADC to measure an additional parameter.

As of this update, the plant sensor prototype is functionally complete.
Calibration may be needed to dial in the soil moisture sensor when it's
used in soil, but this will be a simple mapping function if it is even
required. The sensor currently outputs 0 for dry air and 1 when
submersed in water.