#13 Reworked Sensor API
The sensor module now exposes a fairly simple set of methods that allow
the user to update all sensors for all modules with one function call,
update individual sensors for a single plant, or update all sensors for
a single plant.

Saving can be done either on a batch update or on request.

This is a re-do of a previous commit.