Close #15: Fixed Light Sensor Errors and Readout
Fixed an issue where the TSL2550D light sensor occasionally refused to
respond to commands intermittently. This specific issue was tied to the
lack of a forced wait time for onboard ADC conversion. The downside of
this fix is that - in order to correctly acquire results - it takes
nearly a second to convert ADC results. A possible solution to this
method would be to use the extended mode, which reduces total wait time
to 160 ms.

The second issue with the light sensor was tied to incorrectly
interpreting the conversion algorithm in the datasheet. This fix seems
to have corrected the readout of the device.