#13 Improved pump support and API
This commit has two main features:

1) The pump is now governed by a list that each control object is
capable of accessing through the parent class. If any of the components
of this list are int(1), the pump will turn on. I may later add support
for different options here, but this would be easiest to interface with
valves if we intend to keep support for them.

2) The API for the control module has been expanded in order to make it
easier to interface with as per issue #13. This largely comes in the
form of the method control(). This method accepts a variety of inputs in
order to provide the user a few different ways to update the controls.