Verified bus abstraction functionality.
Everything worked great aside from a typo. I've also manually resolved the changes in a recent commit to the master that I would normally have waited on, but it was actually a fatal error and was preventing me from running the test scripts. The additions pertain to finishing the migration of the analog sensor power control code into the SensorCluster class.