Added CAD files for greenhouse structure
These files serve as a preliminary sketch as to the dimensions and features of the greenhouse. The light fixtures have no means of mounting in these files, as a support beam can be added at a later point and is not integral to the cutting that will be done in the workshop.