#23 Remapped IO pins to work with hardware.
The IO pins have been remapped to account for the fact that only 1
control object is going to be used per greenhouse box.

The test script has also been updated to work with the new sense and
control models.

Along with this commit, I have also verified that the production control
circuitry is working properly and is ready to be mounted inside the
enclosure.