Close #9. Fixed ADC and verified that sensor is working.
The sensor appears to be working and the pi is able to get the water
level within about a quarter of a centimeter.

We will need to add the water level sensor to a tube of some sort in
order to keep it rigid and exposed to water.

The ADC changes should also be reflected  in the soil moisture sensor.
This change can be made during calibration.