#25 Re-calibrated depth sensor and scaled soil moisture sensor
Offset for depth sensor was too large and has been decreased to more
accurately reflect the values. It appears that the small measurements
threw off the calculated slope (or the slight nonlinearity meant that it
went off target. Additionally, the size of the container was too large
in the file and has been lowered.

The soil moisture now represents itself as a value from 0 (being dry
air) to about 1 when submerged in water. Values in between will
represent a percentage of moisture content. It appears that values will
be around the 30-70% mark for adequate soil conditions for most plants.